Final Additions to Project and Mastery assignment
Made final additions to group project using diversity indices to make conclusions about Hosack and Colden Flora, and I would like to this to also count as my mastery assignment (vegan skills from portfolio 2). I also mapped a lot of the species in order to get a better idea of the global context of Hosack's and Colden's Flora.